requirements.txt
PyYAMLgitpythonrequestsollamaopenpyxlpymongopandas